I modified the way the dates are displayed as labels on the graph.  I noticed that the graph could really only handle a maximum 6 or 7 labels (dates) before it too crowded to read, so I made a little formula that reduces the number of labels displayed on the graph, but still displays the labels corresponding to each weight.
It first calculates at what spacing it needs to display six weights on
the graph (eg. if there are 6 weights, it needs a spacing of 1, if
there are 12 weights, it needs a spacing of 2).  Then it iterates
through each of the weight/date pairs, checks to see if the key is
divisible by that spacing. If it is divisible, then it will display
that label, if it isn't, it moves on to the next label.