Made the link to the installation manual a local link and added the installation instructions to the project.
Fixed a bug in how the main controller added up # bats for each room.

Moved the links around and added a link to weekend care for the menubar.

Added how many mealworms are eaten by the colony each day to the colony status page.

Made a lot of changes to the weekend care page.

Made some more room for the menu by reducing the left and right margins by 1 cm.