We have a new menu!
For some reason the div-session is not a part of the div-menubar, so the gray area cuts off right before the end of the session info, so I commented out the gray area.  Also, there's some weird clipping that happens when you scroll over the menu.  Probably the two css files are conflicting.