Added a listing of bats to the user summary page.
Added a column for last weigh date for the cage list and the bat list.  Shows the date as red when it's been more than a week and the bat weights are monitored.

Added a link to the user list and new user from the login page.  Removed the link to the animal care staff from the sidebar.